requirements.txt
requests
dnspython
telnetlib3
pydantic
